debugging app.py added instances and migrations
